chore: refactor dashboard data module and update imports
