--- bc_obps/common/api/v1/__init__.py
@@ -1,2 +1,0 @@
-# ruff: noqa: F401
-from .dashboard_data import tiles
