+++ bc_obps/common/schema/__init__.py
@@ -1,1 +1,2 @@
 # ruff: noqa
+from .dashboard_data import *
