--- bc_obps/common/schema/v1/__init__.py
@@ -1,2 +1,0 @@
-# ruff: noqa
-from .dashboard_data import *
