+++ bc_obps/service/data_access_service/dashboard_service.py
@@ -1,13 +1,10 @@
+import logging
 from typing import Optional
 from uuid import UUID
-
 from django.db.models import QuerySet
 from django.db.models import Q
 from common.models import DashboardData
-from common.schema.v1 import DashboardDataSchemaOut
-
-import logging
-
+from common.schema import DashboardDataSchemaOut
 from service.data_access_service.user_service import UserDataAccessService
 
 logger = logging.getLogger(__name__)
